requirements.txt
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.4
django-extensions==2.1.4
djangorestframework==3.9.0
idna==2.5
kombu==4.2.2
newsapi-python==0.2.3
numpy==1.15.4
pandas==0.23.4
psycopg2==2.7.6.1
Pygments==2.3.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.17.1
six==1.12.0
url-normalize==1.4.1
urllib3==1.21.1
vine==1.1.4
watson-developer-cloud==2.5.1
websocket-client==0.48.0
